# backend/tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swings API",
    "description": "Swings public + admin HTTP API. Mutating endpoints are annotated; GETs are still fully callable but not listed here (see FDN-02 scope).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to deploying host"
    }
  ],
  "paths": {
...
    "/api/admin/members/{id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "delete_member",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member deleted"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/members/{id}/billing-portal": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_member_billing_portal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingPortalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stripe billing portal URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPortalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Member has no subscription"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/admin/members/{id}/subscription/cancel": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_member_subscription_cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription scheduled to cancel at period end"
          },
          "400": {
            "description": "Member has no subscription"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/members/{id}/subscription/resume": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_member_subscription_resume",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Member id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription cancellation reversed"
          },
          "400": {
            "description": "Member has no subscription"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/member/billing-portal": {
      "post": {
        "tags": [
          "member"
        ],
        "operationId": "post_billing_portal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingPortalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stripe billing portal session URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPortalResponse"
                }
              }
            }
          },
          "400": {
            "description": "No subscription on file"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/member/courses/{course_id}/progress": {
      "put": {
        "tags": [
          "member"
        ],
        "operationId": "update_progress",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "Course identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Enrollment progress updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseEnrollment"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/member/profile": {
      "put": {
        "tags": [
          "member"
        ],
        "operationId": "update_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/member/subscription/cancel": {
      "post": {
        "tags": [
          "member"
        ],
        "operationId": "post_subscription_cancel",
        "responses": {
          "200": {
            "description": "Subscription scheduled to cancel at period end"
          },
          "400": {
            "description": "No subscription on file"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/member/subscription/resume": {
      "post": {
        "tags": [
          "member"
        ],
        "operationId": "post_subscription_resume",
        "responses": {
          "200": {
            "description": "Subscription cancellation reversed"
          },
          "400": {
            "description": "No subscription on file"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingPortalRequest": {
        "type": "object",
        "properties": {
          "return_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BillingPortalResponse": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
...
      "CourseEnrollment": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "course_id",
          "progress",
          "enrolled_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "course_id": {
            "type": "string"
          },
          "progress": {
            "type": "integer",
            "format": "int32"
          },
          "enrolled_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ProgressUpdate": {
        "type": "object",
        "required": [
          "progress"
        ],
        "properties": {
          "progress": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "type": [
              "string",
              "null"
            ]
          },
          "website_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "twitter_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "linkedin_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "youtube_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "instagram_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "type": [
              "string",
              "null"
            ]
          },
          "website_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "twitter_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "linkedin_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "youtube_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "instagram_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "Member",
          "Admin"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}